* src/base/ftobject.c, src/base/ftsynth.c, src/base/ftstroker.c,         src/bdf/bdfdrivr.c: removed compiler warnings

+++ src/bdf/bdfdrivr.c
@@ -308,11 +308,11 @@ THE SOFTWARE.
 
       prop = bdf_get_font_property( font, (char *)"AVERAGE_WIDTH" );
       if ( ( prop != NULL ) && ( prop->value.int32 >= 10 ) )
-        root->available_sizes->width =  prop->value.int32 / 10;
+        root->available_sizes->width =  (short)( prop->value.int32 / 10 );
 
       prop = bdf_get_font_property( font, (char *)"PIXEL_SIZE" );
       if ( prop != NULL )
-        root->available_sizes->height = prop->value.int32;
+        root->available_sizes->height = (short) prop->value.int32;
       else
       {
         prop = bdf_get_font_property( font, (char *)"POINT_SIZE" );
